Clean up - Refactor resource controller - Part 2
