+++ incubator/virtualcluster/pkg/syncer/resources/node/util.go
