+++ incubator/virtualcluster/pkg/syncer/resources/pod/controller.go
@@ -21,13 +21,8 @@ import (
 	"time"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	coreinformers "k8s.io/client-go/informers/core/v1"
-	clientset "k8s.io/client-go/kubernetes"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
 	listersv1 "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
@@ -38,24 +33,24 @@ import (
 	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/conversion"
 	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/manager"
 	mc "github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/mccontroller"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/metrics"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/reconciler"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/utils"
 )
 
 type controller struct {
-	client                    v1core.CoreV1Interface
-	multiClusterPodController *mc.MultiClusterController
-	informer                  coreinformers.Interface
-
-	workers       int
-	queue         workqueue.RateLimitingInterface
+	// super master pod client
+	client v1core.CoreV1Interface
+	// super master informer/listers/synced functions
+	informer      coreinformers.Interface
 	podLister     listersv1.PodLister
 	podSynced     cache.InformerSynced
 	serviceSynced cache.InformerSynced
 	nsLister      listersv1.NamespaceLister
 	nsSynced      cache.InformerSynced
-
+	// Connect to all tenant master pod informers
+	multiClusterPodController *mc.MultiClusterController
+	// UWS queue
+	workers int
+	queue   workqueue.RateLimitingInterface
+	// Checker timer
 	periodCheckerPeriod time.Duration
 }
 
@@ -146,188 +141,6 @@ func (c *controller) enqueuePod(obj interface{}) {
 	c.queue.Add(key)
 }
 
-func (c *controller) StartDWS(stopCh <-chan struct{}) error {
-	return c.multiClusterPodController.Start(stopCh)
-}
-
-func (c *controller) Reconcile(request reconciler.Request) (reconciler.Result, error) {
-	klog.Infof("reconcile pod %s/%s %s event for cluster %s", request.Namespace, request.Name, request.Event, request.Cluster.Name)
-
-	var operation string
-	switch request.Event {
-	case reconciler.AddEvent:
-		operation = "pod_add"
-		defer recordOperation(operation, time.Now())
-		err := c.reconcilePodCreate(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.Pod))
-		recordError(operation, err)
-		if err != nil {
-			klog.Errorf("failed reconcile pod %s/%s CREATE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	case reconciler.UpdateEvent:
-		operation = "pod_update"
-		defer recordOperation(operation, time.Now())
-		err := c.reconcilePodUpdate(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.Pod))
-		recordError(operation, err)
-		if err != nil {
-			klog.Errorf("failed reconcile pod %s/%s UPDATE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	case reconciler.DeleteEvent:
-		operation = "pod_delete"
-		defer recordOperation(operation, time.Now())
-		err := c.reconcilePodRemove(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.Pod))
-		recordError(operation, err)
-		if err != nil {
-			klog.Errorf("failed reconcile pod %s/%s DELETE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	}
-	return reconciler.Result{}, nil
-}
-
-func (c *controller) reconcilePodCreate(cluster, namespace, name string, vPod *v1.Pod) error {
-	// load deleting pod, don't create any pod on super master.
-	if vPod.DeletionTimestamp != nil {
-		return c.reconcilePodUpdate(cluster, namespace, name, vPod)
-	}
-
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, namespace)
-	_, err := c.podLister.Pods(targetNamespace).Get(name)
-	if err == nil {
-		return c.reconcilePodUpdate(cluster, namespace, name, vPod)
-	}
-
-	newObj, err := conversion.BuildMetadata(cluster, targetNamespace, vPod)
-	if err != nil {
-		return err
-	}
-
-	pPod := newObj.(*v1.Pod)
-
-	// check if the secret in super master is ready
-	// we must create pod after sync the secret.
-	saName := "default"
-	if pPod.Spec.ServiceAccountName != "" {
-		saName = pPod.Spec.ServiceAccountName
-	}
-
-	pSecret, err := utils.GetSecret(c.client, targetNamespace, saName)
-	if err != nil {
-		return fmt.Errorf("failed to get secret: %v", err)
-	}
-
-	if pSecret.Labels[constants.SyncStatusKey] != constants.SyncStatusReady {
-		return fmt.Errorf("secret for pod is not ready")
-	}
-
-	var client *clientset.Clientset
-	tenantCluster := c.multiClusterPodController.GetCluster(cluster)
-	if tenantCluster == nil {
-		klog.Infof("cluster %s is gone", cluster)
-		return nil
-	}
-	client, err = tenantCluster.GetClient()
-	if err != nil {
-		return err
-	}
-	vSecret, err := utils.GetSecret(client.CoreV1(), namespace, saName)
-	if err != nil {
-		return fmt.Errorf("failed to get secret: %v", err)
-	}
-
-	services, err := c.getPodRelatedServices(cluster, pPod)
-	if err != nil {
-		return fmt.Errorf("failed to list services from cluster %s cache: %v", cluster, err)
-	}
-
-	if len(services) == 0 {
-		return fmt.Errorf("service is not ready")
-	}
-
-	conversion.MutatePod(vPod, pPod, vSecret, pSecret, services)
-
-	_, err = c.client.Pods(targetNamespace).Create(pPod)
-	if errors.IsAlreadyExists(err) {
-		klog.Infof("pod %s/%s of cluster %s already exist in super master", namespace, name, cluster)
-		return nil
-	}
-	return err
-}
-
-func (c *controller) getPodRelatedServices(cluster string, pPod *v1.Pod) ([]*v1.Service, error) {
-	var services []*v1.Service
-	list, err := c.informer.Services().Lister().Services(conversion.ToSuperMasterNamespace(cluster, metav1.NamespaceDefault)).List(labels.Everything())
-	if err != nil {
-		return nil, err
-	}
-	services = append(services, list...)
-
-	list, err = c.informer.Services().Lister().Services(pPod.Namespace).List(labels.Everything())
-	if err != nil {
-		return nil, err
-	}
-	services = append(services, list...)
-
-	return services, nil
-}
-
-func (c *controller) reconcilePodUpdate(cluster, namespace, name string, vPod *v1.Pod) error {
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, namespace)
-	pPod, err := c.podLister.Pods(targetNamespace).Get(name)
-	if err != nil {
-		if errors.IsNotFound(err) {
-			// if the pod on super master has been deleted and syncer has not
-			// deleted virtual pod with 0 grace period second successfully.
-			// we depends on periodic check to do gc.
-			return nil
-		}
-		return err
-	}
-
-	if vPod.DeletionTimestamp != nil {
-		if pPod.DeletionTimestamp != nil {
-			// pPod is under deletion, waiting for UWS bock populate the pod status.
-			return nil
-		}
-		deleteOptions := metav1.NewDeleteOptions(*vPod.DeletionGracePeriodSeconds)
-		deleteOptions.Preconditions = metav1.NewUIDPreconditions(string(pPod.UID))
-		err = c.client.Pods(targetNamespace).Delete(name, deleteOptions)
-		if errors.IsNotFound(err) {
-			return nil
-		}
-		return err
-	}
-
-	updatedPod := conversion.CheckPodEquality(pPod, vPod)
-	if updatedPod != nil {
-		pPod, err = c.client.Pods(targetNamespace).Update(updatedPod)
-		if err != nil {
-			return err
-		}
-	}
-
-	// pod has been updated by tenant controller
-	if !equality.Semantic.DeepEqual(vPod.Status, pPod.Status) {
-		c.enqueuePod(pPod)
-	}
-
-	return nil
-}
-
-func (c *controller) reconcilePodRemove(cluster, namespace, name string, vPod *v1.Pod) error {
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, namespace)
-	opts := &metav1.DeleteOptions{
-		PropagationPolicy: &constants.DefaultDeletionPolicy,
-	}
-	err := c.client.Pods(targetNamespace).Delete(name, opts)
-	if errors.IsNotFound(err) {
-		klog.Warningf("pod %s/%s of cluster (%s) is not found in super master", namespace, name, cluster)
-		return nil
-	}
-	return err
-}
-
 func (c *controller) AddCluster(cluster mc.ClusterInterface) {
 	klog.Infof("tenant-masters-pod-controller watch cluster %s for pod resource", cluster.GetClusterName())
 	err := c.multiClusterPodController.WatchClusterResource(cluster, mc.WatchOptions{})
@@ -344,15 +157,4 @@ func (c *controller) RemoveCluster(cluster mc.ClusterInterface) {
 // assignedPod selects pods that are assigned (scheduled and running).
 func assignedPod(pod *v1.Pod) bool {
 	return len(pod.Spec.NodeName) != 0
-}
-
-func recordOperation(operation string, start time.Time) {
-	metrics.PodOperations.WithLabelValues(operation).Inc()
-	metrics.PodOperationsDuration.WithLabelValues(operation).Observe(metrics.SinceInSeconds(start))
-}
-
-func recordError(operation string, err error) {
-	if err != nil {
-		metrics.PodOperationsErrors.WithLabelValues(operation).Inc()
-	}
 }
